requirements.txt
Flask
Pillow
torch
img
os
glob
torchvision
numpy
sklearn
imblearn
cv2
tqdm
threading
queue
shutil
uuid
concurrent.futures
gradio